// frontend-react/src/types/stats/ComparisonSummary.ts
export interface ComparisonSummary {
    currentMonthExpenses: number;
    previousMonthExpenses: number;
    currentMonthIncomes: number;
    previousMonthIncomes: number;

    currentYearExpenses: number;
    previousYearExpenses: number;
    currentYearIncomes: number;
    previousYearIncomes: number;
}